requirements.txt
tensorflow>=2.8.0
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=8.0.0
joblib>=1.0.0
opencv-python>=4.5.0
tqdm>=4.60.0
pathlib2>=2.3.0; python_version < '3.4'